+++ searchlib/src/vespa/searchlib/datastore/buffer_type.h
@@ -32,10 +32,10 @@ protected:
 public:
     class CleanContext {
     private:
-        uint64_t &_extraBytes;
+        size_t &_extraBytes;
     public:
-        CleanContext(uint64_t &extraBytes) : _extraBytes(extraBytes) {}
-        void extraBytesCleaned(uint64_t value);
+        CleanContext(size_t &extraBytes) : _extraBytes(extraBytes) {}
+        void extraBytesCleaned(size_t value);
     };
     
     BufferTypeBase(const BufferTypeBase &rhs) = delete;
@@ -53,7 +53,7 @@ public:
     // Initialize reserved elements at start of buffer.
     virtual void initializeReservedElements(void *buffer, size_t reservedElements) = 0;
     virtual size_t elementSize() const = 0;
-    virtual void cleanHold(void *buffer, uint64_t offset, uint64_t numElems, CleanContext cleanCtx) = 0;
+    virtual void cleanHold(void *buffer, size_t offset, size_t numElems, CleanContext cleanCtx) = 0;
     size_t getArraySize() const { return _arraySize; }
     void flushLastUsed();
     virtual void onActive(uint32_t bufferId, size_t *usedElems, size_t &deadElems, void *buffer);
@@ -91,7 +91,7 @@ public:
     void destroyElements(void *buffer, size_t numElems) override;
     void fallbackCopy(void *newBuffer, const void *oldBuffer, size_t numElems) override;
     void initializeReservedElements(void *buffer, size_t reservedElements) override;
-    void cleanHold(void *buffer, uint64_t offset, uint64_t numElems, CleanContext cleanCxt) override;
+    void cleanHold(void *buffer, size_t offset, size_t numElems, CleanContext cleanCxt) override;
     size_t elementSize() const override { return sizeof(EntryType); }
 };
 
@@ -150,7 +150,7 @@ BufferType<EntryType>::initializeReservedElements(void *buffer, size_t reservedE
 
 template <typename EntryType>
 void
-BufferType<EntryType>::cleanHold(void *buffer, uint64_t offset, uint64_t numElems, CleanContext)
+BufferType<EntryType>::cleanHold(void *buffer, size_t offset, size_t numElems, CleanContext)
 {
     EntryType *e = static_cast<EntryType *>(buffer) + offset;
     for (size_t j = numElems; j != 0; --j) {
